% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStaticDownload.R
\name{downloadSpecialNaturalEarthData}
\alias{downloadSpecialNaturalEarthData}
\title{Download static data}
\usage{
downloadSpecialNaturalEarthData(delete_zip = TRUE)
}
\arguments{
\item{delete_zip}{boolean. Defaults to TRUE. Whether to delete zip folder
after download}
}
\value{
None. Output is written to
\code{rappdirs::user_data_dir("PhotoBioDynamics")}.
}
\description{
Download and process shapefiles for plotting.
This may take a while, depending on your connection.
}
\examples{
\dontrun{

 downloadSpecialNaturalEarthData(check_valid = TRUE, delete_zip = TRUE)
}

}
