% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataModelWrangling.R
\name{getInputData}
\alias{getInputData}
\title{Get input data}
\usage{
getInputData(year, delete.originaldata = TRUE)
}
\arguments{
\item{year}{integer. Define a reference year for recycled data (fpar/lai)
end for the time series of outputs to be written.}

\item{delete.originaldata}{logical. Delete the original/raw data that could
likely comprises a large amount of memory. Defaults to \code{TRUE}}
}
\value{
None. Optional raw output is written to
\code{rappdirs::user_data_dir("PhotoBioDynamics")}, processed model input into
"./data/input".
}
\description{
Download and process required input data that is open access and currently
(year 2020) available.
This may take a while, depending on your connection.
Processed model input data is stored at "./data/input
}
\examples{
\dontrun{

 getInputData(year = 2000)
}

}
