% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelRunner.R
\name{modelRunner}
\alias{modelRunner}
\title{A model wrapper to run different models}
\usage{
modelRunner(
  path_output = NULL,
  outvar = c("GPP", "NPP"),
  format = c("GLOBAL_TS", "BIOME_TS", "GRID"),
  paramfile = NULL
)
}
\arguments{
\item{path_output}{character. define a path to write output. Defaults to
\code{rappdirs::user_data_dir("PhotoBioDynamics")}}

\item{outvar}{single character or vector. Define the output variables,
choose between gross primary production - "GPP", net primary prodcution
\itemize{
\item "NPP" and/or net ecosytem production "NEP
}}

\item{format}{single character or vector. Choose between raster time series
for the defined \code{"outvar"} - "GRID", a global sum time series -
"GLOBAL_TS" and/or a summed time series for each of 14 biomes ("BIOME_TS).}

\item{paramfile}{string. Defaults to \code{NULL}. Provide a filepath to a parameter
file.}
}
\value{
None
}
\description{
Simulation of C3 photosynthesis in terms of Gross Primary Production (GPP),
as well as Net Primary Production (NPP) and Net Ecosystem Production (NEP)
based on a modified Farquhar model (Schaphoff et al. 2019, Haxeltine and
Prentice 1996, Farquhar et al. 1980).
This is a vectorized global gridd approach that uses directly and currently
(year 2020) available remote sensing products (open access).
}
\examples{
\dontrun{

 modelRunner(outvar = c("GPP", "NPP"), format = c("GLOBAL_TS", "BIOME_TS"))
}

}
