% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{wranglingClimateData}
\alias{wranglingClimateData}
\title{Climate data wrangler}
\usage{
wranglingClimateData(year, delete.originaldata = TRUE)
}
\arguments{
\item{year}{integer. Define a reference year for recycled data (fpar/lai)
end for the time series of outputs to be written.}

\item{delete.originaldata}{logical. Delete the original/raw data that could
likely comprises a large amount of memory. Defaults to \code{TRUE}}
}
\value{
None
}
\description{
Download and process required climate data that is open access and currently
(year 2020) available.
Processed data is written to the package's data folder.
}
\examples{
\dontrun{

 wranglingClimateData(year = 2000)
}

}
