% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataModelWrangling.R
\name{getInputData}
\alias{getInputData}
\title{Get input data}
\usage{
getInputData(latestyear, timerange = 30, delete.originaldata = TRUE)
}
\arguments{
\item{latestyear}{integer. Define the latest year for the input data.
LAI/FPAR and PAR are averaged and recycled since their long-term variations
play a minor role within this scope.}

\item{timerange}{integer. Define the timerange to be used for Temperature and
CO2 data. Defaults to \code{30}}

\item{delete.originaldata}{logical. Delete the original/raw data that could
likely comprises a large amount of memory. Defaults to \code{TRUE}}
}
\value{
None. Output is written to
\code{rappdirs::user_data_dir("PhotoBioDynamics")}
}
\description{
Download and process required input data that is open access and currently
(year 2020) available.
This may take a while, depending on your connection.
}
\examples{
\dontrun{

 getInputData(year = 2000)
}

}
